- runs on Mac - using intptr_t instead of int, you ensure that the type can accommodate the size of a pointer on the target platform. This is a safer way to perform pointer-to-integer conversions

+++ Engine/src/Platform/OpenGL/VertexArrayOpenGL.cpp
@@ -58,7 +58,7 @@ namespace Engine
         {
             ENGINE_WARN("glEnableVertexAttribArray({0})", index);
             glEnableVertexAttribArray(index);
-			ENGINE_WARN("glVertexAttribPointer( {0},{1}, {2},{3}, {4},{5})", index, element.GetElementCount(), ShaderTypeToOpenGLType(element.type), (int)element.normalized ? GL_TRUE : GL_FALSE, layout.GetStride(), (int)((char*)NULL + element.offset));
+			ENGINE_WARN("glVertexAttribPointer( {0},{1}, {2},{3}, {4},{5})", index, element.GetElementCount(), ShaderTypeToOpenGLType(element.type), (int)element.normalized ? GL_TRUE : GL_FALSE, layout.GetStride(), (int)((intptr_t)(char*)NULL + element.offset));
 
             glVertexAttribPointer(
                 index,
